refactor: 뷰 컴포넌트 패키지 개선
상품 관리는 `product` 패키지, 주문 관리는 `order` 패키지에서 관리함.
