+++ frontend/src/component/product/UpdateProductAmountDialog.vue
